# airbyte-connector-builder-server/connector_builder/generated/models/known_exception_info.py
# coding: utf-8
#
# Copyright (c) 2022 Airbyte, Inc., all rights reserved.
#

from __future__ import annotations

import re  # noqa: F401
from datetime import date, datetime  # noqa: F401
from typing import Any, Dict, List, Optional  # noqa: F401

from pydantic import AnyUrl, BaseModel, EmailStr, validator  # noqa: F401


class KnownExceptionInfo(BaseModel):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.

    KnownExceptionInfo - a model defined in OpenAPI

        message: The message of this KnownExceptionInfo.
        exception_class_name: The exception_class_name of this KnownExceptionInfo [Optional].
        exception_stack: The exception_stack of this KnownExceptionInfo [Optional].
    """

    message: str
    exception_class_name: Optional[str] = None
    exception_stack: Optional[List[str]] = None


KnownExceptionInfo.update_forward_refs()