# airbyte-connector-builder-server/connector_builder/generated/models/streams_list_request_body.py
# coding: utf-8
#
# Copyright (c) 2022 Airbyte, Inc., all rights reserved.
#

from __future__ import annotations

import re  # noqa: F401
from datetime import date, datetime  # noqa: F401
from typing import Any, Dict, List, Optional  # noqa: F401

from pydantic import AnyUrl, BaseModel, EmailStr, validator  # noqa: F401


class StreamsListRequestBody(BaseModel):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.

    StreamsListRequestBody - a model defined in OpenAPI

        manifest: The manifest of this StreamsListRequestBody.
        config: The config of this StreamsListRequestBody.
    """

    manifest: Dict[str, Any]
    config: Dict[str, Any]


StreamsListRequestBody.update_forward_refs()