# airbyte-connector-builder-server/src/main/openapi/openapi.yaml
openapi: 3.0.0
info:
  description: |
    Connector Builder Server API

  version: "1.0.0"
  title: Connector Builder Server API
  contact:
    email: contact@airbyte.io
  license:
    name: MIT
    url: "https://opensource.org/licenses/MIT"
externalDocs:
  description: Find out more about Connector Builder
  url: "https://docs.airbyte.com/connector-development/config-based/overview/"

paths:
  /v1/stream/read:
    post:
      summary: Reads a specific stream in the source. TODO in a later phase - only read a single slice of data.
      operationId: readStream
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StreamReadRequestBody"
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreamRead"
        "400":
          $ref: "#/components/responses/ExceptionResponse"
        "422":
          $ref: "#/components/responses/InvalidInputResponse"
  /v1/streams/list:
    post:
      summary: List all streams present in the connector manifest, along with their specific request URLs
      operationId: listStreams
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StreamsListRequestBody"
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreamsListRead"
        "400":
          $ref: "#/components/responses/ExceptionResponse"
        "422":
          $ref: "#/components/responses/InvalidInputResponse"
  /v1/manifest_template:
    get:
      summary: Return a connector manifest template to use as the default value for the yaml editor
      operationId: getManifestTemplate
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: string
                description: Connector manifest template string

components:
  schemas:
    StreamRead:
      type: object
      required:
        - logs
        - slices
      properties:
        logs:
          type: array
          description: The LOG AirbyteMessages that were emitted during the read of this slice
          items:
            type: object
            # $ref: "#/components/schemas/AirbyteProtocol/definitions/AirbyteLogMessage"
        slices:
          type: array
          description: The stream slices returned from the read command. If no stream slicer is configured, this should contain a single item containing all of the results.
          items:
            type: object
            required:
              - pages
            properties:
              pages:
                type: array
                description: The pages returned from the read command. If no pagination is configured, this should contain a single item containing all of the results.
                items:
                  type: object
                  required:
                    - records
                  properties:
                    records:
                      type: array
                      description: The RECORD AirbyteMessages coming from the read operation for this page
                      items:
                        type: object
                        # $ref: "#/components/schemas/AirbyteProtocol/definitions/AirbyteRecordMessage"
                    request:
                      $ref: "#/components/schemas/HttpRequest"
                    response:
                      $ref: "#/components/schemas/HttpResponse"
              sliceDescriptor:
                type: object
                description: "An object describing the current slice. This can be omitted if a stream slicer is not configured."
                properties:
                  startDatetime:
                    type: string
                    format: date-time
                  listItem:
                    type: string
              state:
                type: object
                description: The STATE AirbyteMessage emitted at the end of this slice. This can be omitted if a stream slicer is not configured.
                # $ref: "#/components/schemas/AirbyteProtocol/definitions/AirbyteStateMessage"
    StreamReadRequestBody:
      type: object
      required:
        - manifest
        - stream
        - config
      properties:
        manifest:
          type: object
          description: The config-based connector manifest contents
          # $ref: "#/components/schemas/ConnectorManifest"
        stream:
          type: string
          description: Name of the stream to read
        config:
          type: object
          description: The config blob containing the user inputs for testing
        state:
          type: object
          description: The AirbyteStateMessage object to use as the starting state for this read
          # $ref: "#/components/schemas/AirbyteProtocol/definitions/AirbyteStateMessage"
        # --- Potential addition for a later phase ---
        # numPages:
        #   type: integer
        #   description: Number of pages to read from the source for each slice
        #   default: 1
    HttpRequest:
      type: object
      required:
        - url
      properties:
        url:
          type: string
          description: URL that the request was sent to
        parameters:
          type: object
          description: The request parameters that were set on the HTTP request, if any
        body:
          type: object
          description: The body of the HTTP request, if present
        headers:
          type: object
          description: The headers of the HTTP request, if any
    HttpResponse:
      type: object
      required:
        - status
      properties:
        status:
          type: integer
          description: The status of the response
        body:
          type: object
          description: The body of the HTTP response, if present
        headers:
          type: object
          description: The headers of the HTTP response, if any
    # --- Commenting out for now since they do not work with our orval openapi client generator ---
    # ConnectorManifest:
    #   $ref: ../../../../airbyte-cdk/python/airbyte_cdk/sources/declarative/config_component_schema.json
    # AirbyteProtocol:
    #   $ref: ../../../../airbyte-protocol/protocol-models/src/main/resources/airbyte_protocol/airbyte_protocol.yaml
    StreamsListRequestBody:
      type: object
      required:
        - manifest
        - config
      properties:
        manifest:
          type: object
          description: The config-based connector manifest contents
          # $ref: "#/components/schemas/ConnectorManifest"
        config:
          type: object
          description: The config blob containing the user inputs for testing
    StreamsListRead:
      type: object
      required:
        - streams
      properties:
        streams:
          type: array
          items:
            type: object
            description: The stream names present in the connector manifest
            required:
              - name
              - url
            properties:
              name:
                type: string
                description: The name of the stream
              url:
                type: string
                description: The URL to which read requests will be made for this stream
              # --- Potential addition for a later phase ---
              # slices:
              #   type: array
              #   description: list of slices that will be retrieved for this stream
              #   items:
              #     type: object

    # The following exception structs were copied from airbyte-api/src/main/openapi/config.yaml
    InvalidInputProperty:
      type: object
      required:
        - propertyPath
      properties:
        propertyPath:
          type: string
        invalidValue:
          type: string
        message:
          type: string
    KnownExceptionInfo:
      type: object
      required:
        - message
      properties:
        message:
          type: string
        exceptionClassName:
          type: string
        exceptionStack:
          type: array
          items:
            type: string
    InvalidInputExceptionInfo:
      type: object
      required:
        - message
        - validationErrors
      properties:
        message:
          type: string
        exceptionClassName:
          type: string
        exceptionStack:
          type: array
          items:
            type: string
        validationErrors:
          type: array
          items:
            $ref: "#/components/schemas/InvalidInputProperty"

  responses:
    InvalidInputResponse:
      description: Input failed validation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/InvalidInputExceptionInfo"
    ExceptionResponse:
      description: Exception occurred; see message for details.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/KnownExceptionInfo"